moved msg_change_validator_weight
